add raw to tsv, mangodb first commit, restructure
