+++ modules/creds.env
@@ -1,0 +1,7 @@
+##
+## MangoDB server connection details
+##
+
+mangosrv_user = ""
+mangosrv_pass = ""
+mangosrv_db = ""
