+++ modules/db_notes.tsv
@@ -1,0 +1,4 @@
+## TSV File to MangoDB
+## VN_20240401_00_00_00.TSV
+##
+## VN NAME, DATE, TIME, GPS_X, GPS_Y, SPEED, ODOMETER
