+++ requirements.txt
@@ -1,3 +1,5 @@
 Flask
 Flask-SocketIO
 gevent
+pymongo
+python-dotenv
